switch ducks patterns initial commit: topdesign, designlist, category got switched
