--- src/actions/ActionTypes.js
@@ -1,15 +1,0 @@
-// Category List
-export const GET_CATEGORY_LIST_SUCCESS = "GET_CATEGORY_LIST_SUCCESS"
-export const GET_CATEGORY_LIST_FAILURE = "GET_CATEGORY_LIST_FAILURE"
-
-// Design List
-export const GET_DESIGN_LIST_SUCCESS = "GET_DESIGN_LIST_SUCCESS"
-export const GET_DESIGN_LIST_FAILURE = "GET_DESIGN_LIST_FAILURE"
-export const GET_DESIGN_LIST_CLEAR = "GET_DESIGN_LIST_CLEAR"
-export const GET_DESIGN_LIST_COUNT_SUCCESS = "GET_DESIGN_LIST_COUNT_SUCCESS"
-export const GET_DESIGN_LIST_COUNT_FAILURE = "GET_DESIGN_LIST_COUNT_FAILTURE"
-
-// Top Design List
-export const GET_TOP_DESIGN_LIST_SUCCESS = "GET_TOP_DESIGN_LIST_SUCCESS"
-export const GET_TOP_DESIGN_LIST_FAILURE = "GET_TOP_DESIGN_LIST_FAILURE"
-export const GET_TOP_DESIGN_LIST_CLEAR = "GET_TOP_DESIGN_LIST_CLEAR"
